Updated to the new wasm-ast library
