+++ wasm-rpc-stubgen/Cargo.toml
@@ -23,7 +23,8 @@ cargo-component-core = "=0.13.2"
 cargo-component = "=0.13.2"
 dir-diff = "0.3.3"
 fs_extra = "1.3.0"
-golem-wasm-ast = "0.3.0"
+#golem-wasm-ast = "0.3.0"
+golem-wasm-ast = { git = "https://github.com/golemcloud/golem-wasm-ast", branch = "analysis-type-features" }
 golem-wasm-rpc = { path = "../wasm-rpc", version = "0.0.0" }
 heck = "0.5.0"
 id-arena = "2.2.1"
