+++ wasm-rpc/Cargo.toml
@@ -21,13 +21,16 @@ async-recursion = { version = "1.1.1", optional = true }
 async-trait = { version = "0.1.77", optional = true }
 bigdecimal = { version = "0.4.5", optional = true }
 bincode = { version = "2.0.0-rc.3", optional = true }
-golem-wasm-ast = { version = "0.3.0", features = [
-    "analysis",
-    "wave",
+#golem-wasm-ast = { version = "0.3.0", features = [
+#    "analysis",
+#    "wave",
+#], optional = true }
+golem-wasm-ast = { git = "https://github.com/golemcloud/golem-wasm-ast", branch = "analysis-type-features", features = [
+    "analysis", "wave", "json", "bincode", "poem_openapi", "protobuf"
 ], optional = true }
-serde = { version = "1.0.203", optional = true }
-serde_json = { version = "1.0.117", optional = true }
-prost = { version = "0.12.6", optional = true }
+serde = { version = "1.0", optional = true }
+serde_json = { version = "1.0", optional = true }
+prost = { version = "0.12", optional = true }
 wasmtime = { version = "=21.0.1", features = [
     "component-model",
 ], optional = true }
@@ -40,6 +43,7 @@ proptest-arbitrary-interop = "0.1.0"
 
 [build-dependencies]
 prost-build = "0.12.6"
+cargo_metadata = "0.18.1"
 
 [features]
 default = ["host"]
@@ -61,7 +65,7 @@ protobuf = ["dep:bincode", "dep:serde", "dep:prost"]
 serde = ["dep:serde"]
 stub = []
 text = ["wasmtime", "dep:wasm-wave"]
-typeinfo = ["dep:golem-wasm-ast"]
+typeinfo = ["dep:golem-wasm-ast", "protobuf"]
 wasmtime = ["dep:wasmtime", "dep:wasmtime-wasi", "dep:async-recursion", "typeinfo"]
