+++ wasm-rpc/build.rs
@@ -1,20 +1,28 @@
+use cargo_metadata::MetadataCommand;
 use std::io::Result;
 
 fn main() -> Result<()> {
+    let wasm_ast_root = find_package_root("golem-wasm-ast");
+
     let mut config = prost_build::Config::new();
+    config.extern_path(".wasm.ast", "::golem_wasm_ast::analysis::protobuf");
     config.type_attribute(".", "#[cfg(feature = \"protobuf\")]");
-    config.type_attribute(
-        ".",
-        "#[derive(bincode::Encode, bincode::Decode, serde::Serialize, serde::Deserialize)]",
-    );
     config.compile_protos(
         &[
-            "proto/wasm/rpc/type.proto",
             "proto/wasm/rpc/val.proto",
             "proto/wasm/rpc/witvalue.proto",
             "proto/wasm/rpc/type_annotated_value.proto",
         ],
-        &["proto/"],
+        &[&format!("{wasm_ast_root}/proto"), &"proto".to_string()],
     )?;
     Ok(())
+}
+
+fn find_package_root(name: &str) -> String {
+    let metadata = MetadataCommand::new()
+        .manifest_path("./Cargo.toml")
+        .exec()
+        .unwrap();
+    let package = metadata.packages.iter().find(|p| p.name == name).unwrap();
+    package.manifest_path.parent().unwrap().to_string()
 }
