+++ wasm-rpc/src/lib.rs
@@ -31,6 +31,11 @@ mod extractor;
 #[cfg(feature = "json")]
 pub mod json;
 
+/// Legacy version of conversion to and from JSON, in the presence of golem-wasm-ast generated type information
+#[cfg(feature = "json")]
+// #[deprecated(since = "0.0.38", note = "Use the JSON representation of `TypeAnnotatedValue` instead")]
+pub mod legacy_json;
+
 /// Protobuf-defined value types and conversion to them
 #[cfg(feature = "protobuf")]
 pub mod protobuf;
@@ -39,12 +44,15 @@ pub mod protobuf;
 #[cfg(feature = "serde")]
 pub mod serde;
 
+/// Conversion to/from the WAVE format
 #[cfg(feature = "text")]
 mod text;
 
+/// A version of values annotated with golem-wasm-ast generated type information
 #[cfg(feature = "typeinfo")]
 mod type_annotated_value;
 
+/// Conversion to/from wasmtime's value representation
 #[cfg(feature = "wasmtime")]
 pub mod wasmtime;
